requirements.txt
streamlit==1.32.0
wikipedia==1.4.0
# openai==1.107.3
fpdf==1.7.2
python-dotenv==1.0.1
groq==0.31.1 
pyttsx3==2.90  
# gTTS==2.3.2
# playsound==1.3.0